==> 04-missing.Rmd <==
# Missing values

```{r, include=FALSE}
knitr::opts_chunk$set(warning = FALSE, message = FALSE, echo = TRUE)
```

```{r, echo=FALSE}
library(tidyverse)
```

## title.bacis
```{r, echo=FALSE}
load('data/title.basics.RData')
```

### Missing values by column
```{r}
colSums(is.na(title.basics)) %>%
  sort(decreasing = TRUE)
```


### Missing patterns 
```{r, echo=FALSE}
title.basics <- rename(title.basics, c('time'='runtimeMinutes', 'tiType'='titleType','primTitle'='primaryTitle','orign'='originalTitle'))
```
```{r}
source("scripts/visna.R")
visna(title.basics)
visna(title.basics, percent=TRUE)
```

From the graph, we can see that there are no missing values in columns `nconst` and `primaryName`. Most missing values are in columns `deathYear` and `birthYear`. This might be due to the difficulty of getting the specific birth and death time and the fact that some people are not dead. Only a minority of missing values are in columns `primaryProfession` and `knownForTitles`. Furthermore, `deathYear` and `birthYear` are both missing in most rows with missing data. It might be due to the high correlation between these 2 columns. Only a very small proportion of rows have no missing values.
