==> 07-conclusion.Rmd <==
# Conclusion

The goal of our project is to find out some interesting things about different movie genres. We firstly focused on ratings and tried to find the most popular movies. During this process we found that the number of votes can have a big influence on your results. You can get very different results when you use different standards. We found that although there are more than 600,000 movies in IMDB database, many of them have very few votes. From the interactive part we see that the number of movies has been increasing dramatically in recent decades, especially `Documentary` movies. Many of them have high ratings but few votes. Using both average rating and number of votes, we define a metric 'number of stars' to find most influential movies and directors. Most of them are very famous. We also looked into titles of different genres and found some interesting results. 

We learned a lot during this project. Firstly we practiced what we have learned this whole semester and created different types of graphs, including the coding techniques in R and theoretical knowledge about data visualization. We also learned how to use javascript on your web page to make animations and interactive components. Mostly importantly, we understood why it is called `Exploratory Data Analysis`: you never know what you will find and what graphs you will draw before you start. It is a process of "exploration", you may find new things and decide to try new techniques at the middle of it. It's different from our usual assignments where we already have the question and requirements about what graphs to plot. 

Of course there are some limitations in our project. Firstly we actually don't have some important variables such as budget, box office and country. We can explore much more interesting things if we have these data. Secondly we can dig more into the reasons behind the results, such as why some genres like `Documentary` has been increasing so fast, are there many student or unprofessional works, or why people have preference in different genres.
